Move collateral icon to icons
