--- v3/components/CollateralIcon/package.json
@@ -1,10 +1,0 @@
-{
-  "name": "@snx-v3/CollateralIcon",
-  "private": true,
-  "main": "index.ts",
-  "version": "0.0.1",
-  "dependencies": {
-    "@chakra-ui/react": "^2.2.8",
-    "@snx-v3/icons": "workspace:*"
-  }
-}
