+++ v3/components/CollateralTypeSelector/package.json
@@ -6,6 +6,7 @@
   "dependencies": {
     "@chakra-ui/icons": "^2.0.9",
     "@chakra-ui/react": "^2.2.8",
+    "@snx-v3/icons": "workspace:*",
     "@snx-v3/useCollateralTypes": "workspace:*"
   }
 }
