+++ v3/components/icons/CollateralIcon/index.ts
