+++ v3/components/icons/index.tsx
@@ -1,3 +1,4 @@
 // Base svgs go in static assets, more dynamic svgs here
 export * from './EthereumIcon';
 export * from './SNXIcon';
+export * from './CollateralIcon';
