+++ v3/ui/package.json
@@ -24,7 +24,6 @@
     "@snx-v3/Amount": "workspace:*",
     "@snx-v3/Balance": "workspace:*",
     "@snx-v3/BorderBox": "workspace:*",
-    "@snx-v3/CollateralIcon": "workspace:*",
     "@snx-v3/CollateralTypeSelector": "workspace:*",
     "@snx-v3/Constants": "workspace:*",
     "@snx-v3/Multistep": "workspace:*",
@@ -32,6 +31,7 @@
     "@snx-v3/TrendText": "workspace:*",
     "@snx-v3/etherscanLink": "workspace:*",
     "@snx-v3/format": "workspace:*",
+    "@snx-v3/icons": "workspace:*",
     "@snx-v3/useAccounts": "workspace:*",
     "@snx-v3/useApprove": "workspace:*",
     "@snx-v3/useBlockchain": "workspace:*",
